More clean up.

+++ Sources/SF2Lib/Entity/Generator/Definition.cpp
@@ -25,13 +25,13 @@ Definition::convertedValueOf(const Amount& amount) const noexcept
 {
   switch (valueKind_) {
     case ValueKind::coarseOffset: return valueOf(amount) * 32768;
-    case ValueKind::signedCents: return Float(valueOf(amount) / 1200.0_F);
+    case ValueKind::signedCents: return valueOf(amount) / 1200.0_F;
 
     case ValueKind::signedCentsBel:
     case ValueKind::unsignedPercent:
     case ValueKind::signedPercent: return valueOf(amount) / 10.0_F;
 
-    case ValueKind::signedFrequencyCents: return Float(DSP::centsToFrequency(valueOf(amount)));
+    case ValueKind::signedFrequencyCents: return DSP::centsToFrequency(valueOf(amount));
     case ValueKind::signedTimeCents: return DSP::centsToSeconds(valueOf(amount));
 
     default: return valueOf(amount);
